Refactoring

+++ application/posts/views.py
@@ -1,16 +1,16 @@
-from flask import Blueprint, flash, render_template, redirect, url_for, request, current_app
+from flask import Blueprint, flash, render_template, redirect, url_for, request, current_app, abort
 from flask_login import login_required, current_user
-from application.posts.forms import PostForm, CommentForm
+from application.posts.forms import PostForm, CommentForm, DeleteForm
 from application.models import Post, Comment, Tag
 from application import db
-from application.decorators import roles_required
+from application.decorators import superuser_required
 
 posts = Blueprint('posts', __name__, template_folder='templates')
 
 
 @posts.route('/create_post', methods=['GET', 'POST'])
 @login_required
-@roles_required('admin', 'writer')
+@superuser_required
 def create_post():
     form = PostForm()
     if form.validate_on_submit():
@@ -30,6 +30,7 @@ def create_post():
 def post(post_id):
     post = Post.query.get_or_404(post_id)
     form = CommentForm()
+    delete_form = DeleteForm()
     if form.validate_on_submit():
         comment = Comment(
             content=form.content.data,
@@ -44,23 +45,26 @@ def post(post_id):
     comments = post.comments.order_by(Comment.timestamp.asc()).paginate(
         page=page, per_page=current_app.config['COMMENTS_PER_PAGE']
     )
-    return render_template('post.html', post=post, form=form, comments=comments)
+    return render_template('post.html', post=post, form=form, comments=comments,
+                           delete_form=delete_form)
 
 
-@posts.route('/post/<int:post_id>/delete_comment/<int:comment_id>', methods=['GET', 'POST'])
+@posts.route('/post/<int:post_id>/<int:comment_id>/delete', methods=['POST'])
 @login_required
 def delete_comment(post_id, comment_id):
     post = Post.query.get_or_404(post_id)
     comment = Comment.query.get_or_404(comment_id)
-    db.session.delete(comment)
-    db.session.commit()
-    flash('Comment has been deleted.', 'success')
-    return redirect(url_for('posts.post', post_id=post.id))
+    if comment.user_id == current_user.id or current_user.is_superuser():
+        db.session.delete(comment)
+        db.session.commit()
+        flash('Comment has been deleted.', 'success')
+        return redirect(url_for('posts.post', post_id=post.id))
+    abort(403)
 
 
 @posts.route('/post/<int:post_id>/update', methods=['GET', 'POST'])
 @login_required
-@roles_required('admin', 'writer')
+@superuser_required
 def update_post(post_id):
     post = Post.query.get_or_404(post_id)
     form = PostForm()
@@ -76,9 +80,9 @@ def update_post(post_id):
     return render_template('create_post.html', form=form, title='Update Post')
 
 
-@posts.route('/post/<int:post_id>/delete', methods=['GET', 'POST'])
+@posts.route('/post/<int:post_id>/delete', methods=['POST'])
 @login_required
-@roles_required('admin', 'writer')
+@superuser_required
 def delete_post(post_id):
     post = Post.query.get_or_404(post_id)
     post_title = post.title
@@ -90,6 +94,10 @@ def delete_post(post_id):
 
 @posts.route('/tag/<tag_name>')
 def tag(tag_name):
+    delete_form = DeleteForm()
+    page = request.args.get('page', 1, type=int)
     tag = Tag.query.filter_by(name=tag_name).first_or_404()
-    posts = Post.query.filter(Post.tags.any(name=tag_name))
-    return render_template('tag.html', tag=tag, posts=posts)
+    tags = Tag.query.order_by(Tag.name)
+    posts = Post.query.filter(Post.tags.any(name=tag_name)).paginate(page=page,
+                                                                per_page=current_app.config['POSTS_PER_PAGE'])
+    return render_template('tag.html', tag=tag, tags=tags, posts=posts, delete_form=delete_form)
